requirements.txt
selenium 
pyautogui
Pillow 
webdriver-manager
opencv-python 
torch 
torchvision 
ultralytics
argparse
pytesseract
